docs: renumber Crawl4AI guide to 08 and switch to a single-file example
